requirements.txt
python-a2s==1.3.0
requests==2.25.1